README_pt_BR
I fully translated the documentation to broaden the reach and accessibility of the project, allowing more people to understand and use it, regardless of language. This translated version facilitates the understanding of the functionalities and applications of co-tracker, making the documentation accessible to a wider audience and promoting its use in different contexts and research areas.